[MRG] Automatic merge, trunk 20189 to 20918
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/experimental@20925 b456876b-0849-0410-b77d-98878d47e9d5

+++ kaltura_integration/db/convertscripts/convertsqls.sh
@@ -7,10 +7,10 @@
 #
 # If you don't have php-cli the, you can use:
 # sh convertsqls.sh domain.com/subdomain/
-#
 
 VERSION="4.0"
 
+# Display help information?
 if [ "$1" = "-h" ]; then
 	echo "Usage: ./convertsqls.sh <host> <tikiversion> or .//convertsqls.sh -h for this help"
 	echo "       where <host> is the virtualhost/root/ for your tiki, IF NOT SET just runs php from the command line instead"
@@ -18,40 +18,48 @@ if [ "$1" = "-h" ]; then
 	exit 0
 fi
 
-TIKISERVER=""
-if [ -z $1]; then 
-TIKISERVER=$1
+if [ ! -f "convertsqls.sh" ]; then
+	echo "Error: this script needs to be executed from its directory, db/convertscripts/"
+	exit 0
 fi
 
+# set tikiserver for remote conversation/download if specified as parameter
+TIKISERVER=""
+if [ -z $1]; then
+  TIKISERVER=$1
+fi
+
+# set version if specified as parameter
 if [ "$2" ] ; then
   VERSION=$2
 fi
 
+# tiki.sql actually is MySQL code (our DB codebase) so just copy it for finished mysql scripts
 cp ../tiki.sql ../tiki-$VERSION-mysql.sql
 cp ../tiki.sql ../tiki-$VERSION-mysqli.sql
-# /* the scripts use mysql.sql */
 
+# do convert
 if [$TIKISERVER = ""]; then
-echo "Local run of php ..."
-php -f mysql3topgsql72.php > pgsql72.sql.tmp
-php -f mysql3tosybase.php > sybase.sql.tmp
-php -f mysql3tosqlite.php > sqlite.sql.tmp
-php -f mysql3tooci8.php > oci8.sql.tmp
+	# convert locally
+	echo "Local run of php ..."
+	php -f mysql_to_pgsql.php version=$VERSION > pgsql.sql.tmp
+	php -f mysql_to_sqlite.php version=$VERSION > sqlite.sql.tmp
 else
-wget -O pgsql72.sql.tmp "http://$TIKISERVER/db/convertscripts/mysql3topgsql72.php?version=$VERSION" 
-wget -O sybase.sql.tmp "http://$TIKISERVER/db/convertscripts/mysql3tosybase.php?version=$VERSION" 
-wget -O sqlite.sql.tmp "http://$TIKISERVER/db/convertscripts/mysql3tosqlite.php?version=$VERSION"
-wget -O oci8.sql.tmp "http://$TIKISERVER/db/convertscripts/mysql3tooci8.php?version=$VERSION" 
+	# convert remotely and download
+	echo "Running remote scripts and downloading script files ..."
+	wget -O pgsql.sql.tmp "http://$TIKISERVER/db/convertscripts/mysql_to_pgsql.php?version=$VERSION"
+	wget -O sqlite.sql.tmp "http://$TIKISERVER/db/convertscripts/mysql_to_sqlite.php?version=$VERSION"
 fi
 
-rm -f *.sql.tmp 
-rm -f ../tiki-$VERSION-pgsql.sql ../tiki-$VERSION-sybase.sql ../tiki-$VERSION-sqlite.sql ../tiki-$VERSION-oci8.sql
+# remove temporary output files (we don't need the output from conversion scripts)
+rm -f *.sql.tmp
+# remove old converted scripts
+rm -f ../tiki-$VERSION-pgsql.sql ../tiki-$VERSION-sqlite.sql
 
-mv $VERSION.to_pgsql72.sql ../tiki-$VERSION-pgsql.sql
-mv $VERSION.to_sybase.sql ../tiki-$VERSION-sybase.sql
+# move the newly converted/created scripts
+mv $VERSION.to_pgsql.sql ../tiki-$VERSION-pgsql.sql
 mv $VERSION.to_sqlite.sql ../tiki-$VERSION-sqlite.sql
-mv $VERSION.to_oci8.sql ../tiki-$VERSION-oci8.sql
-
 
 echo "Done."
+
 exit 0
